sql-rest-api/pyproject.toml
[project]
name = "sql-rest-api"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "requests>=2.32.3",
    "pyjwt>=2.9.0",
    "cryptography>=43.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sql_rest_api"]
